+++ src/plugins/citation.js
@@ -11,7 +11,7 @@ export default (schema) => {
 			const newTransaction = newState.tr;
 			newState.doc.nodesBetween(0, newState.doc.nodeSize - 2, (node, nodePos) => {
 				if (node.type.name === 'citation') {
-					const key = `${node.attrs.html}-${node.attrs.unstructuredValue}`;
+					const key = `${node.attrs.value}-${node.attrs.unstructuredValue}`;
 					const existingCount = counts[key] && counts[key].count;
 					const nextCount = Object.keys(counts).length + 1;
 					if (existingCount && node.attrs.count !== existingCount) {
@@ -47,61 +47,9 @@ export default (schema) => {
 				return true;
 			});
 
-			/* Check all CitationList nodes to make sure they are updated if */
-			/* didUpdate is true, or if the list is empty, but counts is not */
-			newState.doc.nodesBetween(0, newState.doc.nodeSize - 2, (node, nodePos) => {
-				if (node.type.name === 'citationList') {
-					/* Test whether the html content of the citation list should be */
-					/* updated due to new html in individual citations */
-					const citationListContentChanged = Object.keys(counts).reduce((prev, curr) => {
-						const currCitationData = counts[curr];
-						const prevCitationData =
-							node.attrs.listItems[currCitationData.count - 1] || {};
-						const prevKey = `${prevCitationData.html}-${prevCitationData.unstructuredValue}`;
-						if (prevKey !== curr) {
-							return true;
-						}
-						return prev;
-					}, false);
-
-					if (
-						node.attrs.listItems.length !== Object.keys(counts).length ||
-						didUpdate ||
-						citationListContentChanged
-					) {
-						const listItems = Object.keys(counts)
-							.sort((foo, bar) => {
-								if (counts[foo].count < counts[bar].count) {
-									return -1;
-								}
-								if (counts[foo].count > counts[bar].count) {
-									return 1;
-								}
-								return 0;
-							})
-							.map((key) => {
-								return {
-									html: counts[key].html,
-									value: counts[key].value,
-									unstructuredValue: counts[key].unstructuredValue,
-									count: counts[key].count,
-								};
-							});
-
-						didUpdate = true;
-						newTransaction.setNodeMarkup(nodePos, null, {
-							...node.attrs,
-							listItems: listItems,
-						});
-						newTransaction.setMeta('citation', true);
-					}
-				}
-				return true;
-			});
 			if (didUpdate && newTransaction.doc.selection) {
 				/* Numbers being updated when html changes causes the node to lose focus. */
 				/* This refocuses the node */
-
 				newTransaction.setSelection(newTransaction.doc.selection);
 			}
 			return didUpdate ? newTransaction : null;