+++ src/schemas/base.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export const baseNodes = {
 	doc: {
 		content: 'block+',
@@ -24,19 +22,11 @@ export const baseNodes = {
 			},
 		],
 		toDOM: (node) => {
-			return ['p', { class: node.attrs.class }, 0];
-		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			const attrs = {};
-			if (node.attrs && node.attrs.class) {
-				attrs.className = node.attrs.class;
-			}
-			const emptyChildren = <br />;
-			return (
-				<p {...attrs} key={node.currIndex}>
-					{children || emptyChildren}
-				</p>
-			);
+			/* On server rendering, the placeholder <br /> tag is not inserted. */
+			/* The following check ensures spacing is maintained when server rendering */
+			const isEmpty = Array.isArray(node.content) && !node.content.length;
+			const children = isEmpty ? ['br'] : 0;
+			return ['p', { class: node.attrs.class }, children];
 		},
 	},
 	blockquote: {
@@ -45,9 +35,6 @@ export const baseNodes = {
 		parseDOM: [{ tag: 'blockquote' }],
 		toDOM: () => {
 			return ['blockquote', 0];
-		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return <blockquote key={node.currIndex}>{children}</blockquote>;
 		},
 	},
 	horizontal_rule: {
@@ -62,9 +49,6 @@ export const baseNodes = {
 					view.state.schema.nodes.horizontal_rule.create(),
 				),
 			);
-		},
-		toStatic: (node) => {
-			return <hr key={node.currIndex} />;
 		},
 	},
 	heading: {
@@ -116,51 +100,6 @@ export const baseNodes = {
 		toDOM: (node) => {
 			return [`h${node.attrs.level}`, { id: node.attrs.id }, 0];
 		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			if (node.attrs.level === 1) {
-				return (
-					<h1 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h1>
-				);
-			}
-			if (node.attrs.level === 2) {
-				return (
-					<h2 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h2>
-				);
-			}
-			if (node.attrs.level === 3) {
-				return (
-					<h3 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h3>
-				);
-			}
-			if (node.attrs.level === 4) {
-				return (
-					<h4 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h4>
-				);
-			}
-			if (node.attrs.level === 5) {
-				return (
-					<h5 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h5>
-				);
-			}
-			if (node.attrs.level === 6) {
-				return (
-					<h6 key={node.currIndex} id={node.attrs.id}>
-						{children}
-					</h6>
-				);
-			}
-			return null;
-		},
 	},
 	ordered_list: {
 		content: 'list_item+',
@@ -177,14 +116,6 @@ export const baseNodes = {
 		toDOM: (node) => {
 			return ['ol', { start: node.attrs.order === 1 ? null : node.attrs.order }, 0];
 		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			const attrs = { start: node.attrs.order === 1 ? null : node.attrs.order };
-			return (
-				<ol key={node.currIndex} {...attrs}>
-					{children}
-				</ol>
-			);
-		},
 	},
 	bullet_list: {
 		content: 'list_item+',
@@ -193,9 +124,6 @@ export const baseNodes = {
 		toDOM: () => {
 			return ['ul', 0];
 		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return <ul key={node.currIndex}>{children}</ul>;
-		},
 	},
 	list_item: {
 		content: 'paragraph block*',
@@ -203,9 +131,6 @@ export const baseNodes = {
 		parseDOM: [{ tag: 'li' }],
 		toDOM: () => {
 			return ['li', 0];
-		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return <li key={node.currIndex}>{children}</li>;
 		},
 	},
 	code_block: {
@@ -216,26 +141,12 @@ export const baseNodes = {
 		toDOM: () => {
 			return ['pre', ['code', 0]];
 		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return (
-				<pre key={node.currIndex}>
-					<code>{children}</code>
-				</pre>
-			);
-		},
 	},
 	text: {
 		inline: true,
 		group: 'inline',
 		toDOM: (node) => {
 			return node.text;
-		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return editorProps.renderStaticMarkup ? (
-				children
-			) : (
-				<span key={node.currIndex}>{children}</span>
-			);
 		},
 	},
 	hard_break: {
@@ -245,22 +156,6 @@ export const baseNodes = {
 		parseDOM: [{ tag: 'br' }],
 		toDOM: () => {
 			return ['br'];
-		},
-		toStatic: (node) => {
-			return <br key={node.currIndex} />;
-		},
-	},
-	none: {
-		// empty schema block
-		/* It's not clear to me that the none schema is used. */
-		/* At the moment, it's not included in the defaultNodes prop */
-		/* by default. */
-		group: 'block',
-		toDOM: () => {
-			return ['span'];
-		},
-		toStatic: (node, options, isSelected, isEditable, editorProps, children) => {
-			return <span key={node.currIndex}>{children}</span>;
 		},
 	},
 };
@@ -278,17 +173,16 @@ export const baseMarks = {
 		toDOM: () => {
 			return ['em'];
 		},
-		toStatic: (mark, children, key) => {
-			return <em key={key}>{children}</em>;
-		},
 	},
 
 	strong: {
 		parseDOM: [
 			{ tag: 'strong' },
-			// This works around a Google Docs misbehavior where
-			// pasted content will be inexplicably wrapped in `<b>`
-			// tags with a font-weight normal.
+			/*
+			This works around a Google Docs misbehavior where
+			pasted content will be inexplicably wrapped in `<b>`
+			tags with a font-weight normal.
+			*/
 			{ tag: 'b', getAttrs: (node) => node.style.fontWeight !== 'normal' && null },
 			{
 				style: 'font-weight',
@@ -297,9 +191,6 @@ export const baseMarks = {
 		],
 		toDOM: () => {
 			return ['strong'];
-		},
-		toStatic: (mark, children, key) => {
-			return <strong key={key}>{children}</strong>;
 		},
 	},
 	link: {
@@ -331,27 +222,11 @@ export const baseMarks = {
 			}
 			return ['a', attrs];
 		},
-		toStatic: (mark, children, key) => {
-			return (
-				<a
-					key={key}
-					href={mark.attrs.href}
-					title={mark.attrs.title}
-					target={mark.attrs.target}
-				>
-					{children}
-				</a>
-			);
-		},
-		toEditable: () => {} /* This is a workaround to make the LinkMenu function within tables */,
 	},
 	sub: {
 		parseDOM: [{ tag: 'sub' }],
 		toDOM: () => {
 			return ['sub'];
-		},
-		toStatic: (mark, children, key) => {
-			return <sub key={key}>{children}</sub>;
 		},
 	},
 	sup: {
@@ -359,26 +234,17 @@ export const baseMarks = {
 		toDOM: () => {
 			return ['sup'];
 		},
-		toStatic: (mark, children, key) => {
-			return <sup key={key}>{children}</sup>;
-		},
 	},
 	strike: {
 		parseDOM: [{ tag: 's' }, { tag: 'strike' }, { tag: 'del' }],
 		toDOM: () => {
 			return ['s'];
 		},
-		toStatic: (mark, children, key) => {
-			return <s key={key}>{children}</s>;
-		},
 	},
 	code: {
 		parseDOM: [{ tag: 'code' }],
 		toDOM: () => {
 			return ['code'];
-		},
-		toStatic: (mark, children, key) => {
-			return <code key={key}>{children}</code>;
 		},
 	},
 };